combine the different Cypress tests into one source file
