--- tests/cypress/e2e/for-teacher_page/class_page/create_accounts_page/add_row.cy.js
@@ -1,12 +1,0 @@
-import {loginForTeacher} from '../../../tools/login/login.js'
-import {createClassAndAddStudents} from '../../../tools/classes/class.js'
-
-
-describe('Is able to add rows to create more accounts', () => {
-  it('Passes', () => {
-    loginForTeacher();
-    createClassAndAddStudents();
-    cy.get('#add_multiple_rows').click();
-    cy.get(':nth-child(6) > #username').should('have.value', '')
-  })
-})
