--- tests/cypress/e2e/for-teacher_page/class_page/create_accounts_page/download_login_credentials.cy.js
@@ -1,11 +1,0 @@
-import {loginForTeacher} from '../../../tools/login/login.js'
-import {createClassAndAddStudents} from '../../../tools/classes/class.js'
-
-
-describe('Is able to download login credentials', () => {
-  it('Passes', () => {
-    loginForTeacher();
-    createClassAndAddStudents();
-    cy.readFile('cypress/downloads/accounts.csv')
-  })
-})
